Move DecriptionUtils to "internal" package
