+++ src/main/java/co/unruly/matchers/function/DescribableFunction.java
@@ -4,7 +4,7 @@ import java.lang.invoke.MethodType;
 import java.lang.invoke.SerializedLambda;
 import java.util.function.Function;
 
-import static co.unruly.matchers.function.DescriptionUtils.withPrefixedArticle;
+import static co.unruly.matchers.internal.DescriptionUtils.withPrefixedArticle;
 
 @FunctionalInterface
 public interface DescribableFunction<T, R> extends Function<T, R>, SerializedLambdaResolvable, SingleArgumentDescribableFunctionalInterface {
