+++ src/main/java/co/unruly/matchers/internal/DescriptionUtils.java
@@ -1,4 +1,4 @@
-package co.unruly.matchers.function;
+package co.unruly.matchers.internal;
 
 import static java.util.Arrays.binarySearch;
