+++ src/main/java/co/unruly/matchers/internal/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * This package and any subpackages contain classes used internally by
+ * the library; it is <em>not</em> part of the public API.
+ */
+package co.unruly.matchers.internal;
